# docs/swagger.yaml
definitions:
  model.IServerReq:
    properties:
      datasetNames:
        items:
          type: string
        type: array
      getFeatureMode:
        type: string
      queryParameter:
        properties:
          attributeFilter:
            type: string
        type: object
    type: object
  model.IServerResp:
    type: object
  model.L21:
    properties:
      id:
        type: string
      修建年度:
        type: string
      养护里程:
        type: string
      可绿化里程:
        type: string
      国道调整前路线编号:
        type: string
      地貌代码:
        type: string
      地貌汉字:
        type: string
      备注:
        type: string
      已绿化里程:
        type: string
      所在行政区划代码:
        type: string
      技术等级:
        type: string
      技术等级代码:
        type: string
      改建年度:
        type: string
      断链类型:
        type: string
      是否一幅高速:
        type: string
      是否城管路段:
        type: string
      是否按干线公路管理接养:
        type: string
      是否断头路段:
        type: string
      最近一次修复养护年度:
        type: string
      止点名称:
        type: string
      止点桩号:
        type: string
      涵洞数量(个):
        type: string
      省际出入口:
        type: string
      管养单位名称:
        type: string
      设计时速(公里/小时):
        type: string
      起点名称:
        type: string
      起点桩号:
        type: string
      路基宽度(米):
        type: string
      路段收费性质:
        type: string
      路线名称:
        type: string
      路线编号:
        type: string
      路面宽度(米):
        type: string
      车道数量(个):
        type: string
      里程(公里):
        type: string
      重复路段线路编号:
        type: string
      重复路段终点桩号:
        type: string
      重复路段起点桩号:
        type: string
      面层厚度(厘米):
        type: string
      面层类型:
        type: string
      面层类型代码:
        type: string
    type: object
  model.ParamLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.ParamSignUp:
    properties:
      password:
        type: string
      repassword:
        type: string
      unit:
        type: string
      username:
        type: string
      verifycode:
        type: string
    required:
    - password
    - repassword
    - unit
    - username
    - verifycode
    type: object
  model.RespUser:
    properties:
      role:
        type: string
      unit:
        type: string
      user_id:
        type: integer
      user_id_str:
        type: string
      username:
        type: string
    type: object
  respcode.ResponseData:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: object
    type: object
info:
  contact: {}
  description: 交通一张图
  title: 交通一张图后端系统
  version: 1.0.0
paths:
...
  /api/v1/data/info/{infotype}/{year}/{level}:
    get:
      consumes:
      - application/json
      description: 获取数据库原始数据接口 访问后会更新缓存
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: '查询类型 : road(路)  bridge(桥) tunnel(隧道) service(服务区) portal(收费门架)
          toll(收费站)'
        in: path
        name: infotype
        required: true
        type: string
      - description: '查询年份 格式: 202X '
        in: path
        name: year
        required: true
        type: string
      - description: '查询等级 : 0(高速) 1(一级) 2(二级) 3(三级) 4(四级) 5(等外)'
        in: path
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新缓存接口
      tags:
      - 数据操作api
    post:
      consumes:
      - application/json
      description: 删除数据库原始数据接口 访问后会删除缓存
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: '查询类型 : road(路)  bridge(桥) tunnel(隧道) service(服务区) portal(收费门架)
          toll(收费站)'
        in: path
        name: infotype
        required: true
        type: string
      - description: '查询年份 格式: 202X '
        in: path
        name: year
        required: true
        type: string
      - description: '查询等级 : 0(高速) 1(一级) 2(二级) 3(三级) 4(四级) 5(等外)'
        in: path
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除指定数据
      tags:
      - 数据操作api
  /api/v1/data/info/{infotype}/{year}/{level}/query:
    post:
      consumes:
      - application/json
      description: 查询表数据（年报）数据接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: '查询类型 : road(路)  bridge(桥) tunnel(隧道) service(服务区) portal(收费门架)
          toll(收费站)'
        in: path
        name: infotype
        required: true
        type: string
      - description: '查询年份 格式: 202X '
        in: path
        name: year
        required: true
        type: string
      - description: '查询等级 : 0(高速) 1(一级) 2(二级) 3(三级) 4(四级) 5(等外)'
        in: path
        name: level
        required: true
        type: integer
      - description: 道路信息（示例）
        in: body
        name: 查询信息
        required: true
        schema:
          $ref: '#/definitions/model.L21'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取查询数据
      tags:
      - 数据操作api
  /api/v1/data/menus:
    get:
      consumes:
      - application/json
      description: 用于前端渲染数据管理侧边栏
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取数据导航栏
      tags:
      - 用户相关api
  /api/v1/home/menus:
    get:
      consumes:
      - application/json
      description: 用于前端渲染用户家目录侧边栏
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取导航栏数据
      tags:
      - 用户相关api
  /api/v1/home/roles:
    get:
      consumes:
      - application/json
      description: 请求后可以拿到权限数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  additionalProperties:
                    items:
                      type: string
                    type: array
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取权限数据接口
      tags:
      - 权限相关api
    head:
      consumes:
      - application/json
      description: 请求后可以拿到用户数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 恢复默认权限接口
      tags:
      - 权限相关api
    post:
      consumes:
      - application/json
      description: 请求后可以跟新权限数据 构造一个新的权限json 作为 body
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  additionalProperties:
                    items:
                      type: string
                    type: array
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新权限数据接口
      tags:
      - 权限相关api
  /api/v1/home/users:
    get:
      consumes:
      - application/json
      description: 请求后可以拿到用户数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/model.RespUser'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户数据接口
      tags:
      - 用户相关api
    post:
      consumes:
      - application/json
      description: 请求后可以将post请求body中提供的用户新数据替换原来的用户数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: 用户信息
        required: true
        schema:
          $ref: '#/definitions/model.RespUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户数据接口
      tags:
      - 用户相关api
  /api/v1/home/users/{id}:
    delete:
      consumes:
      - application/json
      description: 请求后可以删除用户数据（直接从数据库删除）
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除用户数据接口
      tags:
      - 用户相关api
  /api/v1/home/users/query:
    get:
      consumes:
      - application/json
      description: 请求后可以获取指定的用户数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 属性
        in: query
        name: column
        required: true
        type: string
      - description: 值
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/model.RespUser'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 查询用户数据接口
      tags:
      - 用户相关api
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 首页登录
      parameters:
      - description: 用户登录信息
        in: body
        name: 登录信息
        required: true
        schema:
          $ref: '#/definitions/model.ParamLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 处理登录请求
      tags:
      - 登陆api
  /api/v1/signup:
    get:
      consumes:
      - application/json
      description: 申请一个注册码 重复请求上一个就会失效 重启后端服务也会失效
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 申请一个注册码
      tags:
      - 注册api
    post:
      consumes:
      - application/json
      description: 首页注册
      parameters:
      - description: 用户注册信息
        in: body
        name: 注册信息
        required: true
        schema:
          $ref: '#/definitions/model.ParamSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 处理注册请求
      tags:
      - 注册api
  /api/v2/iserver/services/{service}/rest/{service}:
    post:
      consumes:
      - application/json
      description: 用于转发IServer数据，并做权限检验
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 查询组建构建信息
        in: body
        name: 查询信息
        required: true
        schema:
          $ref: '#/definitions/model.IServerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IServerResp'
      security:
      - ApiKeyAuth: []
      summary: 获取IServer数据接口
      tags:
      - IServer相关api
swagger: "2.0"